完善客户管理。商品管理。供应商管理。进货管理模块

+++ jmjs/src/com/controller/PurchaseController.java
@@ -114,7 +114,7 @@ public class PurchaseController extends HttpServlet{
 			value = Integer.valueOf(goodsPrice);
 		}
 		
-		purchase.setGoodsPrice((Integer)value);
+		purchase.setGoodsPrice((Double)value);
 		
 		
 		String goodsNumber = req.getParameter("goodsNumber");
@@ -189,7 +189,7 @@ public class PurchaseController extends HttpServlet{
 			value = Integer.valueOf(goodsPrice);
 		}
 		
-		purchase.setGoodsPrice((Integer)value);
+		purchase.setGoodsPrice((Double)value);
 		
 		
 		String goodsNumber = req.getParameter("goodsNumber");
